web-app/api/requirements.txt
# MPAC Web App — FastAPI backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0
bcrypt>=4.0
pyjwt>=2.8
anthropic>=0.40.0
websockets>=14.0
python-multipart>=0.0.9
